data-collector/requirements.txt
fastapi
uvicorn
polars
pymongo
python-dotenv
pytest
yfinance
redis